+++ src/site/markdown/release-notes.md
@@ -1,5 +1,5 @@
 <!--
-  Copyright 2017 BloomReach Inc (https://www.bloomreach.com)
+  Copyright 2017-2019 BloomReach Inc (https://www.bloomreach.com)
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -16,11 +16,19 @@
 
 ## Version Compatibility
 
-| Hippo CMS | Gallery Background Processor |
-| --------- |------------------------------| 
-| 11.2.3    | 1.0.x                        |
+| Bloomreach CMS | Gallery Background Processor |
+| -------------- |------------------------------| 
+| 13.0.0         | 2.x                          |
+| 11.2.3         | 1.x                          |
 
 ## Release Notes
+
+### 2.0.0  
+
+<p class="smallinfo">Release date: 4 July 2019</p>
+
++ Upgrade to BloomReach Experience version 13
++ Rename artifact groupId's and Java packages from `org.onhippo` to `org.bloomreach`
 
 ### 1.0.0